Added Bus Interface Opcodes
I added the wb and rb opcodes, which write to and read from buses.
They only work with hard-coded bus numbers for the moment.  I think
that will probably be sufficient for the expected use cases, but I can
revisit it if I find a good use case for run-time bus indexing.