Added Arithmetic Opcodes
The VM now has add, sub, mul, div and sdiv (signed division), which
should make it essentially usable.  I'll probably add some bitwise
opcodes and then move on to working on the assembler.  I also included
a small bugfix for the labeling of the VM's debug output in this diff.