Added (Hopefully) Final Opcodes
I added comparison operators as well as a missing binary not.  At this
point, I think the VM is ready for use.  I also added some convenience
functions for converting to the binary true and false values the VM
uses and generating random test values in opcodes_test.go.  At some
point I'd like to retrofit all of the repetitive tests for arithmetic
and bitwise operators to use pseudo-random data as well.