Added Comments and Opcode/Label Lines to Assembler
I made two changes to the assembler internals.  The first is a change
to the lexer to add support for comments (# comments out everything
until the end of the line) by simply lexing a # followed by everything
up to the next new line as a single newline token.

The second was a change to the parser to allow a label on the same
line as anything else that can form a valid line, an opcode or a
%words or %org directive.