Changes to vmtest, Added Benchmark
I modified vmtest to add a verbose flag, which will display a debug
output after each clock cycle, and a memory size argument.  I also
added a benchmark in the VM tests and removed the auto-reset mechanism
in VM's Clock() method, which was consuming a ton of time each cycle.